backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
boto3==1.34.24
pydantic-settings==2.2.1
python-dotenv==1.0.0
APScheduler==3.10.4
